+++ src/main/java/spring/app/model/Company.java
@@ -45,21 +45,10 @@ public class Company {
             inverseJoinColumns = {@JoinColumn(name = "genre_id")})
     private Set<Genre> bannedGenres;
 
-
-    @OneToMany(mappedBy = "company")
-    private SongThatPlayNow songThatPayNow;
-
-
     @OneToMany(mappedBy = "company")
     private Set<SongQueue> songQueues;
 
-    public SongThatPlayNow getSongThatPayNow() {
-        return songThatPayNow;
-    }
 
-    public void setSongThatPayNow(SongThatPlayNow songThatPayNow) {
-        this.songThatPayNow = songThatPayNow;
-    }
 
     public Company(String name, LocalTime startTime, LocalTime closeTime, User user, OrgType orgType) {
         this.name = name;
